Created a new list and added comments
Added a large amount of comments to the code so I know which parts do what. Also created a new list which stores the names of the genres only. I had an issue where the listbox for the tracks was showing the genre name then the tracks so I needed to split this up.