Genre Lists Working
Realised I didn't need the for loop which loops through the tracks, I removed this and now I have each genre in it's own list with all the tracks. Going to look at how I can display the lists so it starts a index 1 as index 0 is the title of the genre. Also going to look at adding a new variable, something along the lines of genreIndex which will be a parameter in my read file function and I can pass this parameter with the function on my buttons (or scroll bar if I manage to work out how to scroll to different genres).