Got half of the Setup Page Working
Genera Title and the Listbox with the current tracks are populating as they should be. The next and previous arrows also work without issues. The code was exactly the same as the code on the main form, I just had to change the function name and edit the code to fit the setup page objects.