Final Commit
Removed test listboxes, neatened up the code a little before submission.

Sadly I was not able to fully implement the Setup menu to it's full capabilities. I also could not work out how to implement the scroll bar so I just have a couple of buttons in the place of the scroll bar for now.